Ensure changeblock reuses serialization from Elastic.Changelog
