+++ src/services/Elastic.Changelog/Creation/ChangelogCreationService.cs
@@ -5,7 +5,7 @@
 using System.IO.Abstractions;
 using Elastic.Changelog.Configuration;
 using Elastic.Changelog.GitHub;
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 using Elastic.Documentation.Configuration;
 using Elastic.Documentation.Diagnostics;
 using Elastic.Documentation.Services;
