+++ src/services/Elastic.Changelog/Creation/ChangelogFileWriter.cs
@@ -5,7 +5,7 @@
 using System.IO.Abstractions;
 using Elastic.Changelog.Configuration;
 using Elastic.Changelog.Serialization;
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 using Elastic.Documentation.Diagnostics;
 using Microsoft.Extensions.Logging;
