+++ src/services/Elastic.Changelog/Creation/CreateChangelogArgumentsValidator.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information
 
 using Elastic.Changelog.Configuration;
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 using Elastic.Documentation.Configuration;
 using Elastic.Documentation.Diagnostics;
