+++ src/services/Elastic.Changelog/Creation/PrInfoProcessor.cs
@@ -4,7 +4,7 @@
 
 using Elastic.Changelog.Configuration;
 using Elastic.Changelog.GitHub;
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 using Elastic.Documentation.Diagnostics;
 using Microsoft.Extensions.Logging;
