+++ src/services/Elastic.Changelog/Rendering/ChangelogRenderUtilities.cs
@@ -2,7 +2,7 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 
 namespace Elastic.Changelog.Rendering;
