+++ src/services/Elastic.Changelog/Rendering/Markdown/IndexMarkdownRenderer.cs
@@ -4,9 +4,9 @@
 
 using System.IO.Abstractions;
 using System.Text;
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 using static System.Globalization.CultureInfo;
-using static Elastic.Documentation.Changelog.ChangelogEntryType;
+using static Elastic.Changelog.ChangelogEntryType;
 
 namespace Elastic.Changelog.Rendering.Markdown;
