+++ src/services/Elastic.Changelog/Serialization/ChangelogMapper.cs
@@ -2,7 +2,7 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-using Elastic.Documentation.Changelog;
+using Elastic.Documentation;
 using Riok.Mapperly.Abstractions;
 
 namespace Elastic.Changelog.Serialization;
